Add: support chrome
