+++ .eslintrc.json
@@ -12,6 +12,7 @@
 		"eslint:recommended"
 	],
   "parserOptions": {
-    "sourceType": "module"
+    "sourceType": "module",
+    "ecmaVersion": "2021"
   }
 }
